Track besøk i router subscription
