+++ index.html
@@ -13,23 +13,17 @@
     
     <link rel="preload" href="https://cdn.nav.no/aksel/fonts/SourceSans3-normal.woff2" as="font" type="font/woff2" crossorigin/>
     <script>
-        window.sendWithEventData = (_event, payload) => {
-            const newUrl = new URL(window.location.href);
-            if (newUrl.host.includes('localhost')) return false;
-
-            newUrl.search = '';
-
-            return {
-                ...payload,
-                url: newUrl.toString()
-            };
+        window.beforeSendAnalytics = (_event, payload) => {
+            const url = new URL(window.location.href);
+            if (url.host.includes('localhost')) return false;
+            return payload
     };
     </script>
 
     <slot not-environment="prod">
         <script
                 defer
-                data-before-send="sendWithEventData"
+                data-before-send="beforeSendAnalytics"
                 src="https://cdn.nav.no/team-researchops/sporing/sporing-dev.js"
                 data-website-id="bc5cfdcd-ecc1-402b-8cb0-c30913f0bd13"
                 data-auto-track="false"
@@ -38,7 +32,7 @@
     <slot environment="prod">
         <script
                 defer
-                data-before-send="sendWithEventData"
+                data-before-send="beforeSendAnalytics"
                 src="https://cdn.nav.no/team-researchops/sporing/sporing.js"
                 data-website-id="a8c2ec9c-2846-4154-8841-5db26494171a"
                 data-auto-track="false"
